==> 1. Converting Yelp Dataset.ipynb <==
# 1. Converting Yelp Dataset

import pandas as pd

### Business table

First I load the business JSON file in a pandas dataframe.

# Load business JSON file. 
business_json_path = 'JSON\yelp_academic_dataset_business.json'
business = pd.read_json(business_json_path, lines=True)
business.categories[0]

I am interested in giving insights to CoffeeKing, a new startup coffee company that wishes to provide a unique and novel experience to its customers. Therefore, the businesses I will need to analyze are coffee shops. I select the businesses that belong to a coffee category.

# Find businesses that belong to a Coffee category to reduce the business table. 
business = business[business['categories'].notna()]
df_b = business[business['categories'].str.contains('Coffee')]
df_b

### Tips table 

Now I will reduce the tip table to have only those meant for coffee shops. 

# Load tip JSON file.
tip_json_path = 'JSON\yelp_academic_dataset_tip.json'
tips = pd.read_json(tip_json_path, lines=True)

# Remove columns from the business table. 
df_b_tmp = df_b[['business_id', 'name']]
# Join the reduced business table with the tip table using business_id.
df_t = df_b_tmp.set_index('business_id').join(tips.set_index('business_id'))
df_t

### Review Table

Both the review and user JSON files are huge, which provoques a memory issue when trying to load them. 
There is a method that allow me to load this datasets into a pandas dataframe.

First, I can reduce memory usage by specifying the data type of each column. The specifications of the table are in the documentation from Yelp.

review_json_path = 'JSON\yelp_academic_dataset_review.json'
size = 1000000
review = pd.read_json(review_json_path, lines=True,
                      dtype={'review_id':str,'user_id':str,
                             'business_id':str,'stars':int,
                             'date':str,'text':str,'useful':int,
                             'funny':int,'cool':int},
                      chunksize=size)

Then I merge the reviews and business tables, using business_id, to have one with the reviews of coffee shops by chunks using the code provided at:

    https://gist.github.com/gyhou/66628a4bfff04b4b67b625173d6ec194#file-review_chunks-py

chunk_list = []
for chunk_review in review:
    # Renaming column name to avoid conflict with business overall star rating
    chunk_review = chunk_review.rename(columns={'stars': 'review_stars'})
    # Inner merge with edited business file so only reviews related to the business remain
    chunk_merged = pd.merge(df_b, chunk_review, on='business_id', how='inner')
    # Show feedback on progress
    print(f"{chunk_merged.shape[0]} out of {size:,} related reviews")
    chunk_list.append(chunk_merged)
# After trimming down the review file, concatenate all relevant data back to one dataframe
df = pd.concat(chunk_list, ignore_index=True, join='outer', axis=0)

The last step is to remove all the columns that belong to the business table. 

drop_columns = ['name','address', 'city', 'state', 'postal_code',
                'latitude', 'longitude', 'stars',
                'review_count', 'is_open', 'attributes',
                'hours','categories']
df_r = df.drop(drop_columns, axis=1)
df_r

### User Table

To build the user table, I perform the same procedure as for the review. The only difference is that now I use the new review table I created at the last step and use the user_id key to join both tables. 

User table specifications: 

users_json_path = 'JSON\yelp_academic_dataset_user.json'
size = 1000000
users = pd.read_json(users_json_path, lines=True,
                      dtype={'user_id':str,'name':str, 'review_count' :int,
                             'yelping_since':str,'friends':str,
                             'useful':int,'funny':int,'cool':int,
                             'fans':int,'elite':str, 'average_stars':float,
                             'compliment_hot':int, 'compliment_more':int,
                             'compliment_profile':int, 'compliment_cute':int,
                             'compliment_list':int, 'compliment_note': int,
                             'compliment_plain':int, 'compliment_cool':int,
                             'compliment_funny':int, 'compliment_writer':int,
                             'compliment_photos':int
                            
                            },
                      chunksize=size)

Join users and reviews by chunk using user_id:

chunk_list = []
for chunk_user in users:
    # Drop columns that aren't needed
    #--chunk_review = chunk_review.drop(['review_id','useful','funny','cool'], axis=1)
    # Renaming column name to avoid conflict with business overall star rating
    chunk_user = chunk_user.rename(columns={'useful': 'user_useful'})
    chunk_user = chunk_user.rename(columns={'cool': 'user_cool'})
    chunk_user = chunk_user.rename(columns={'funny': 'user_funny'})
    # Inner merge with edited business file so only reviews related to the business remain
    chunk_merged = pd.merge(df_r, chunk_user, on='user_id', how='inner')
    # Show feedback on progress
    print(f"{chunk_merged.shape[0]} out of {size:,} related reviews")
    chunk_list.append(chunk_merged)
# After trimming down the review file, concatenate all relevant data back to one dataframe
df_tmp = pd.concat(chunk_list, ignore_index=True, join='outer', axis=0)

Drop columns from the review table: 

drop_columns = ['review_id','business_id', 'review_stars', 'useful', 'funny',
                'cool', 'text', 'date']

df_u = df_tmp.drop(drop_columns, axis=1)
df_u

Finally, I save our resulting tables in csv files. 

csv_name = "yelp_business.csv"
df_b.to_csv(csv_name, index=False)
csv_name = "yelp_tip.csv"
df_t.to_csv(csv_name, index=False)
csv_name = "yelp_review.csv"
df_r.to_csv(csv_name, index=False)
csv_name = "yelp_user.csv"
df_u.to_csv(csv_name, index=False)
