==> 2. Cleaning dataset.ipynb <==
# 2. Cleaning Dataset

import pandas as pd
import datetime
from pandasql import sqldf
pd.set_option('display.max_colwidth', 1000)
pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', None)

First I load the tables I converted from Yelp dataset.

coffee_business = pd.read_csv("Converted\yelp_business.csv")
coffee_tip = pd.read_csv("Converted\yelp_tip.csv")
coffee_review = pd.read_csv("Converted\yelp_review.csv")
coffee_user = pd.read_csv("Converted\yelp_user.csv")

Before working with the datasets, I verify all the data types of the attributes.

coffee_review.info()

I saw that the dates do not have the correct type. Before working with the dataset, I modify every feature with a date to DateTime. 

coffee_tip['date'] = pd.to_datetime(coffee_tip['date'])
coffee_review['date'] = pd.to_datetime(coffee_review['date'])
coffee_user['yelping_since'] = pd.to_datetime(coffee_user['yelping_since'])

coffee_review.info()

The hour's column in the business table contains the opening and closing hours of each coffee shop by day of the week. To make the dataset tidy, I will create a new table using this attribute called hours. That table will have 14 attributes, one for the opening hour and the closing hour of each day. 

# The function receives a string and verifies that the hour has the correct format. 
# The time format is changed to an int as hh.mm
def timeFormat(hour):
    for i in range(len(hour)):
        if i == 0 or i == 1:
            if not hour[i].isnumeric():
                hour = '0' + hour
    h = datetime.datetime.strptime(hour,'%H:%M').time()
    return h.hour + h.minute/100  

# The function receives a string with key: value pairs where the key is the day of the week and the value an hour.
# It returns the opening and closing hours ordered by the day of the week.
def transformToWeekDays(hours):
    week = []
    for day in hours.split(','):
        l = day.split()
        n = len(l[1])
        if l[1][n-1] == '}': 
            n = n-1
        hours = l[1][1:n-1]
        hours = hours.split('-')
        for hour in hours:
            h = timeFormat(hour)
            week.append(h)
    return week

# Construction of the hours table. 
# Format of the hour hh.mm
coffee_weekdays = []
for index, row in coffee_business.iterrows():
    if type(row['hours']) is float or row['is_open'] == 0: 
        coffee_weekdays.append([None, None, None, None, None, None, None])
    else:
        coffee_weekdays.append(transformToWeekDays(row['hours']))

coffee_hours = pd.DataFrame(coffee_business['business_id'])
coffee_hours = coffee_hours.join(pd.DataFrame(coffee_weekdays, index=coffee_business.index, 
                          columns = ["Monday_open", "Monday_close", "Tuesday_open", "Tuesday_close",
                                     "Wednesday_open", "Wednesday_close", "Thursday_open", "Thursday_close",
                                     "Friday_open", "Friday_close", "Saturday_open", "Saturday_close",
                                     "Sunday_open", "Sunday_close"]))

coffee_hours

Now I will remove the attributes and hours columns that will not be necessary for the analysis. 

del coffee_business['attributes']
del coffee_business['hours']
coffee_business_cleaned = coffee_business
coffee_business_cleaned.head()

For the hour analysis, I need businesses that have all their schedule for the week. Therefore, I remove those with null values and that their open hour is the same as their closing hour for all days of the week. 

coffee_hours_cleaned  = sqldf("""
                    SELECT 
                        * 
                    FROM 
                        coffee_hours
                    WHERE 
                        Monday_open IS NOT NULL AND 
                        Tuesday_open IS NOT NULL AND 
                        Wednesday_open IS NOT NULL AND 
                        Thursday_open IS NOT NULL AND 
                        Friday_open IS NOT NULL AND 
                        Saturday_open IS NOT NULL AND 
                        Sunday_open IS NOT NULL
                    AND
                        (Monday_open <> Monday_close OR Tuesday_open <> Tuesday_close OR
                        Wednesday_open <> Wednesday_close OR Thursday_open <> Thursday_close OR
                        Friday_open <> Friday_close OR Saturday_open <> Saturday_close OR
                        Sunday_open <> Sunday_close) 
                        
                    """)
coffee_hours_cleaned.head()

At last, I compare the original size of the hour table with the cleaned one and save all in new CSV files. 

print("          Original   Cleaned")
print("Hours:     ", len(coffee_hours),"     ", len(coffee_hours_cleaned))

csv_name = "yelp_business_cleaned.csv"
coffee_business_cleaned.to_csv(csv_name, index=False)
csv_name = "yelp_tip_cleaned.csv"
coffee_tip.to_csv(csv_name, index=False)
csv_name = "yelp_review_cleaned.csv"
coffee_review.to_csv(csv_name, index=False)
csv_name = "yelp_user_cleaned.csv"
coffee_user.to_csv(csv_name, index=False)
csv_name = "yelp_hours_cleaned.csv"
coffee_hours_cleaned.to_csv(csv_name, index=False)
